pyproject.toml
[tool.poetry]
name = "rag-llm"
version = "0.1.0"
description = ""
authors = ["Nguyen Tuan Dat <nguyentuandat2k3cmg@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.1"
langchain-google-vertexai = "^2.0.8"
langchain-core = "^0.3.24"
langchain = "^0.3.11"
langchain-postgres = "^0.0.12"
psycopg = {extras = ["binary"], version = "^3.2.3"}
langchain-google-genai = "^2.0.6"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
pymongo = "^4.10.1"
langgraph = "^0.2.59"
structlog = "^24.4.0"
langchain-openai = "^0.2.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
